+++ extension.js
@@ -9,20 +9,40 @@ const BUY_ME_A_COFFEE = {
 };
 
 let nbOfTranslations = 0;
+const recentlyUsed = [];
 
 function activate(context) {
 	let disposable = vscode.commands.registerCommand('extension.translateText', function() {
 	const editor = vscode.window.activeTextEditor;
 	const { document, selections } = editor;
-	vscode.window.showQuickPick(languages.map(l => l.name))
+
+	const newLanguagesArray = recentlyUsed.concat(languages);
+	// Add the buy me a coffee link if user use the translator more than 3 times
+	if (nbOfTranslations >= 3) {
+		newLanguagesArray.splice(0, 0, BUY_ME_A_COFFEE);
+	}
+
+	vscode.window.showQuickPick(newLanguagesArray.map(l => l.name))
 		.then(res => {
 			if (!res) return;
-			const { value } = languages.find(t => t.name === res);
-			if (value === BUY_ME_A_COFFEE.value) {
+			const selectedLanguage = newLanguagesArray.find(t => t.name === res);
+			if (selectedLanguage.value === BUY_ME_A_COFFEE.value) {
 				// Buy me a coffee
 				vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(BUY_ME_A_COFFEE.link))
 			} else {
-				// Translate
+				if (recentlyUsed.find(r => r.value === selectedLanguage.value)) {
+					// Remove the recently used language from the list
+					const index = recentlyUsed.findIndex(r => r.value === selectedLanguage.value);
+					recentlyUsed.splice(index, 1);
+				}
+				if (languages.find(r => r.value === selectedLanguage.value)) {
+					// Remove the recently used language from languages list
+					const index = languages.findIndex(r => r.value === selectedLanguage.value);
+					languages.splice(index, 1);
+				}
+				// Add the language in recently used languages
+				recentlyUsed.splice(0, 0, selectedLanguage);
+
 				const promiseResult = selections.map(selection => {
 					const charRange = new vscode.Range(
 						selection.start.line,
@@ -35,7 +55,7 @@ function activate(context) {
 						translate({
 							text: selectedText,
 							source: 'auto',
-							target: value,
+							target: selectedLanguage.value,
 						}, result => {
 							if (!!result && !!result.translation) {
 								resolve({
@@ -51,12 +71,6 @@ function activate(context) {
 				Promise.all(promiseResult)
 					.then(function(results) {
 						nbOfTranslations += 1;
-
-						// Add the buy me a coffee link if user use the translator more than 3 times
-						if (nbOfTranslations === 3) {
-							languages.splice(0, 0, BUY_ME_A_COFFEE);
-						}
-
 						editor.edit(builder => {
 							results.forEach(r => {
 								if (!!r.translation) {